+++ tgui/packages/tgui/interfaces/AntagInfoBrainwashed.tsx
@@ -1,7 +1,7 @@
-import { useBackend } from '../../backend';
-import { Box, Icon, Section, Stack } from '../../components';
-import { Window } from '../../layouts';
-import { ObjectivePrintout, Objective } from './Objectives';
+import { useBackend } from '../backend';
+import { Box, Icon, Section, Stack } from '../components';
+import { Window } from '../layouts';
+import { ObjectivePrintout, Objective } from './common/Objectives';
 
 type Data = {
   objectives: Objective[];
