+++ tgui/packages/tgui/interfaces/AntagInfoGeneric.tsx
@@ -1,24 +1,23 @@
-import { useBackend } from '../../backend';
-import { Section, Stack } from '../../components';
-import { Window } from '../../layouts';
-import { ObjectivePrintout, Objective } from './Objectives';
+import { useBackend } from '../backend';
+import { Section, Stack } from '../components';
+import { Window } from '../layouts';
+import { ObjectivePrintout, Objective } from './common/Objectives';
 
 type Info = {
   antag_name: string;
   objectives: Objective[];
-  brothers: string;
 };
 
-export const AntagInfoBrother = (props, context) => {
+export const AntagInfoGeneric = (props, context) => {
   const { data } = useBackend<Info>(context);
-  const { antag_name, brothers, objectives } = data;
+  const { antag_name, objectives } = data;
   return (
     <Window width={620} height={250}>
       <Window.Content>
         <Section scrollable fill>
           <Stack vertical>
             <Stack.Item textColor="red" fontSize="20px">
-              You are the {antag_name} of {brothers}!
+              You are the {antag_name}!
             </Stack.Item>
             <Stack.Item>
               <ObjectivePrintout objectives={objectives} />
