+++ tgui/packages/tgui/interfaces/AntagInfoSentient.js
@@ -1,6 +1,6 @@
-import { useBackend } from '../../backend';
-import { BlockQuote, Section, Stack } from '../../components';
-import { Window } from '../../layouts';
+import { useBackend } from '../backend';
+import { BlockQuote, Section, Stack } from '../components';
+import { Window } from '../layouts';
 
 export const AntagInfoSentient = (props, context) => {
   const { act, data } = useBackend(context);
