+++ tgui/packages/tgui/interfaces/AntagInfoSeparatist.tsx
@@ -1,7 +1,7 @@
-import { useBackend } from '../../backend';
-import { Icon, Section, Stack } from '../../components';
-import { Window } from '../../layouts';
-import { ObjectivePrintout, Objective } from './Objectives';
+import { useBackend } from '../backend';
+import { Icon, Section, Stack } from '../components';
+import { Window } from '../layouts';
+import { ObjectivePrintout, Objective } from './common/Objectives';
 
 type Info = {
   objectives: Objective[];
